make delete transaction one url and view

+++ crm_app/views/views.py
@@ -91,21 +91,15 @@ def DeleteCustomer(request, pk):
         messages.success(request, "You must be logged in to delete data.")
         return redirect('crm_app:customers')
     
-def DeleteTransactionModal(request, pk):
-    if request.user.is_authenticated:
-        transaction = Transaction.objects.get(id=pk)
-        return render(request, 'crm/partials/delete_transaction.html', {'transaction': transaction})
-    else:
-        messages.success(request, "You must be logged in to delete data.")
-        return redirect('crm_app:home')
-    
 def DeleteTransaction(request, pk):
     if request.user.is_authenticated:
         transaction = Transaction.objects.get(id=pk)
-        messages.success(request, "Transaction has been deleted")
-        transaction.delete()
-        transactions = Transaction.objects.all()
-        return render(request, 'crm/transactions.html', {'transactions': transactions})
+        if request.method == "DELETE":
+            messages.success(request, "Transaction has been deleted")
+            transaction.delete()
+            transactions = Transaction.objects.all() 
+            return render(request, 'crm/transactions.html', {'transactions': transactions})    
+        return render(request, 'crm/partials/delete_transaction.html', {'transaction': transaction})
     else:
         messages.success(request, "You must be logged in to delete data.")
         return redirect('crm_app:home')
